requirements.txt
#
# KeukaSensor Production Requirements
# Compatible with both Python 3.9 (Raspberry Pi) and Python 3.13+ (development)
#

# Web Framework
flask==2.3.3; python_version < "3.10"
flask==3.1.1; python_version >= "3.10"

# Flask dependencies with Python version compatibility
blinker==1.9.0
click>=8.1.8,<8.2.0
colorama==0.4.6
itsdangerous==2.1.2; python_version < "3.10"
itsdangerous==2.2.0; python_version >= "3.10"
jinja2==3.1.3; python_version < "3.10"
jinja2==3.1.6; python_version >= "3.10"
markupsafe==2.1.5; python_version < "3.10"
markupsafe==3.0.2; python_version >= "3.10"
werkzeug==2.3.7; python_version < "3.10"
werkzeug==3.1.3; python_version >= "3.10"

# WebSocket Support
Flask-SocketIO>=5.3.0
eventlet>=0.33.0
python-socketio>=5.8.0

# WSGI Server
gunicorn>=21.0,<22

# SSH Terminal Support
paramiko>=3.0.0

# Computer Vision (OpenCV)
# Pi-optimized headless version for ARM, full version for development
opencv-python-headless==4.6.0.66; platform_machine == "armv6l" or platform_machine == "armv7l" or platform_machine == "aarch64"
opencv-python==4.12.0.88; platform_machine != "armv6l" and platform_machine != "armv7l" and platform_machine != "aarch64"

# Numerical Computing
numpy==1.26.4; python_version < "3.10"
numpy==2.2.6; python_version >= "3.10"

# Hardware Sensors (Raspberry Pi specific, optional on other platforms)
w1thermsensor==2.3.0
RPi.GPIO>=0.7.1; platform_machine == "armv6l" or platform_machine == "armv7l" or platform_machine == "aarch64"

# Development and Testing Dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-flask>=1.2.0
requests>=2.28.0
requests-mock>=1.10.0
coverage>=7.0.0